+++ package.json
@@ -45,9 +45,11 @@
     "@mui/icons-material": "^6.1.5",
     "@mui/material": "^6.1.5",
     "@qubic-lib/qubic-ts-library": "^0.1.4",
+    "jotai": "^2.10.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-iframe": "^1.8.5"
+    "react-iframe": "^1.8.5",
+    "react-router-dom": "^7.0.1"
   },
   "devDependencies": {
     "@types/node": "^22.10.1",
@@ -55,6 +57,7 @@
     "@vitejs/plugin-react": "^4.3.3",
     "electron": "^33.0.2",
     "electron-builder": "^25.1.8",
+    "typescript": "^5.7.2",
     "vite": "^5.4.10"
   }
 }
