+++ src/constants.ts
@@ -1,0 +1,5 @@
+const TICK_OFFSET = 5;
+const BASE_URL = 'https://rpc.qubic.org';
+const POLLING_INTERVAL = 5000;
+
+export { TICK_OFFSET, BASE_URL, POLLING_INTERVAL };
