+++ src/router/index.tsx
@@ -1,0 +1,16 @@
+import Home from '@/pages/home';
+import Login from '@/pages/login';
+import { createBrowserRouter } from 'react-router-dom';
+
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <Home />,
+  },
+  {
+    path: '/login',
+    element: <Login />,
+  },
+]);
+
+export default router;
