+++ src/store/theme.ts
@@ -1,0 +1,7 @@
+import { atom } from 'jotai';
+
+type Theme = 'light' | 'dark';
+
+const themeAtom = atom<Theme>('light');
+
+export default themeAtom;
