+++ src/utils/getTheme.ts
@@ -1,0 +1,37 @@
+import { createTheme } from '@mui/material';
+
+export const getTheme = (mode: 'light' | 'dark') =>
+  createTheme({
+    palette: {
+      mode,
+      ...(mode === 'dark' && {
+        background: {
+          default: '#121928',
+          paper: '#1a2235',
+        },
+      }),
+    },
+    typography: {
+      fontFamily: 'Exo, sans-serif',
+      h6: {
+        fontSize: '1.1rem',
+        fontWeight: 600,
+      },
+      body1: {
+        fontSize: '0.95rem',
+      },
+      button: {
+        textTransform: 'none',
+        fontSize: '0.9rem',
+      },
+    },
+    components: {
+      MuiTableCell: {
+        styleOverrides: {
+          root: {
+            fontSize: '0.9rem',
+          },
+        },
+      },
+    },
+  });
