+++ vite.config.js
@@ -6,5 +6,10 @@ export default defineConfig({
   server: {
     port: 5173, // Default Vite port
   },
+  resolve: {
+    alias: {
+      '@': '/src'
+    }
+  }
 });
